+++ drivers/mac/driver_darwin.go
@@ -2,12 +2,6 @@
 
 package mac
 
-/*
-#include "driver.h"
-#include "bridge.h"
-*/
-import "C"
-
 import (
 	"context"
 	"crypto/md5"
